Move the add licences and add site buttons to a new component
