+++ client/jetpack-cloud/sections/agency-dashboard/sites-overview/index.tsx
@@ -1,5 +1,4 @@
-import { isEnabled } from '@automattic/calypso-config';
-import { Button, Count, WordPressLogo } from '@automattic/components';
+import { Button, Count } from '@automattic/components';
 import { isWithinBreakpoint } from '@automattic/viewport';
 import { useMobileBreakpoint } from '@automattic/viewport-react';
 import { getQueryArg, removeQueryArgs, addQueryArgs } from '@wordpress/url';
@@ -9,13 +8,10 @@ import page from 'page';
 import { useContext, useEffect, useState, useMemo, createRef } from 'react';
 import DocumentHead from 'calypso/components/data/document-head';
 import QueryProductsList from 'calypso/components/data/query-products-list';
-import JetpackLogo from 'calypso/components/jetpack-logo';
-import PopoverMenuItem from 'calypso/components/popover-menu/item';
 import SectionNav from 'calypso/components/section-nav';
 import NavItem from 'calypso/components/section-nav/item';
 import NavTabs from 'calypso/components/section-nav/tabs';
 import SidebarNavigation from 'calypso/components/sidebar-navigation';
-import SplitButton from 'calypso/components/split-button';
 import useFetchDashboardSites from 'calypso/data/agency-dashboard/use-fetch-dashboard-sites';
 import useFetchMonitorVerfiedContacts from 'calypso/data/agency-dashboard/use-fetch-monitor-verified-contacts';
 import { useDispatch, useSelector } from 'calypso/state';
@@ -38,6 +34,7 @@ import SiteContent from './site-content';
 import useDashboardShowLargeScreen from './site-content/hooks/use-dashboard-show-large-screen';
 import SiteContentHeader from './site-content-header';
 import SiteSearchFilterContainer from './site-search-filter-container/SiteSearchFilterContainer';
+import SiteTopHeaderButtons from './site-top-header-buttons';
 import type { Site } from '../sites-overview/types';
 
 import './style.scss';
@@ -199,94 +196,6 @@ export default function SitesOverview() {
 		} );
 	}, [ selectedLicensesSiteId, selectedLicenses ] );
 
-	const AddSiteIssueLicenseButtons = () => {
-		const dispatch = useDispatch();
-		const translate = useTranslate();
-
-		const isWPCOMAtomicSiteCreationEnabled = isEnabled(
-			'jetpack/pro-dashboard-wpcom-atomic-hosting'
-		);
-
-		return (
-			<div
-				className={ classNames( 'sites-overview__add-site-issue-license-buttons', {
-					'is-with-split-button': isWPCOMAtomicSiteCreationEnabled,
-				} ) }
-			>
-				<Button
-					className="sites-overview__issue-license-button"
-					href="/partner-portal/issue-license"
-					onClick={ () =>
-						dispatch(
-							recordTracksEvent( 'calypso_jetpack_agency_dashboard_issue_license_button_click' )
-						)
-					}
-				>
-					{ translate( 'Issue License', { context: 'button label' } ) }
-				</Button>
-
-				{ ! isWPCOMAtomicSiteCreationEnabled && (
-					<Button
-						className="sites-overview__issue-license-button"
-						primary
-						href="https://wordpress.com/jetpack/connect"
-						onClick={ () =>
-							dispatch(
-								recordTracksEvent( 'calypso_jetpack_agency_dashboard_add_site_button_click' )
-							)
-						}
-					>
-						{ translate( 'Add New Site', { context: 'button label' } ) }
-					</Button>
-				) }
-
-				{ isWPCOMAtomicSiteCreationEnabled && (
-					<SplitButton
-						primary
-						whiteSeparator
-						label={ isMobile ? undefined : translate( 'Add new site' ) }
-						onClick={ () =>
-							dispatch(
-								recordTracksEvent(
-									'calypso_jetpack_agency_dashboard_create_wpcom_atomic_site_button_click'
-								)
-							)
-						}
-						href="/partner-portal/create-site"
-						toggleIcon={ isMobile ? 'plus' : undefined }
-					>
-						<PopoverMenuItem
-							onClick={ () => {
-								recordTracksEvent(
-									'calypso_jetpack_agency_dashboard_create_wpcom_atomic_site_button_click'
-								);
-							} }
-							href="/partner-portal/create-site"
-						>
-							<WordPressLogo className="gridicon" size={ 18 } />
-							<span>{ translate( 'Create a new WordPress.com site' ) }</span>
-						</PopoverMenuItem>
-
-						<PopoverMenuItem
-							onClick={ () =>
-								dispatch(
-									recordTracksEvent(
-										'calypso_jetpack_agency_dashboard_connect_jetpack_site_button_click'
-									)
-								)
-							}
-							href="https://wordpress.com/jetpack/connect"
-							isExternalLink
-						>
-							<JetpackLogo className="gridicon" size={ 18 } />
-							<span>{ translate( 'Connect a site to Jetpack' ) }</span>
-						</PopoverMenuItem>
-					</SplitButton>
-				) }
-			</div>
-		);
-	};
-
 	const renderIssueLicenseButton = () => {
 		return (
 			<div className="sites-overview__licenses-buttons">
@@ -344,7 +253,7 @@ export default function SitesOverview() {
 								( selectedLicensesCount > 0 ? (
 									renderIssueLicenseButton()
 								) : (
-									<AddSiteIssueLicenseButtons />
+									<SiteTopHeaderButtons />
 								) )
 							}
 							pageTitle={ pageTitle }
@@ -354,7 +263,7 @@ export default function SitesOverview() {
 
 						{
 							// Render the add site and issue license buttons on mobile as a different component.
-							! isLargeScreen && <AddSiteIssueLicenseButtons />
+							! isLargeScreen && <SiteTopHeaderButtons />
 						}
 						<SectionNav
 							applyUpdatedStyles
